Remove deprecated Elasticsearch() options (#2840)
(cherry picked from commit 20e09c303f14319a6fb085823fe72b9aa21905ac)
