+++ test_elasticsearch/test_types/aliased_types.py
@@ -34,9 +34,7 @@ es = Elasticsearch(
     [{"host": "localhost", "port": 9443}],
     transport_class=Transport,
     sniff_on_start=True,
-    sniffer_timeout=0.1,
     sniff_timeout=1,
-    sniff_on_connection_fail=False,
     max_retries=1,
     retry_on_status={100, 400, 503},
     retry_on_timeout=True,
@@ -103,9 +101,7 @@ def reindex_types() -> None:
 es2 = AsyncElasticsearch(
     [{"host": "localhost", "port": 9443}],
     sniff_on_start=True,
-    sniffer_timeout=0.1,
     sniff_timeout=1,
-    sniff_on_connection_fail=False,
     max_retries=1,
     retry_on_status={100, 400, 503},
     retry_on_timeout=True,
