+++ test_elasticsearch/test_types/sync_types.py
@@ -23,9 +23,7 @@ from elasticsearch.helpers import bulk, reindex, scan, streaming_bulk
 es = Elasticsearch(
     [{"host": "localhost", "port": 9443}],
     sniff_on_start=True,
-    sniffer_timeout=0.1,
     sniff_timeout=1,
-    sniff_on_connection_fail=False,
     max_retries=1,
     retry_on_status={100, 400, 503},
     retry_on_timeout=True,
